backend/app/pyproject.toml
[project]
name = "llm-game"
version = "0.1.0"
description = "Backend for an AI-powered text adventure game"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.124.4",
    "langchain-groq>=1.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "dirty-equals>=0.11",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
